Add support of Session Manager for Rest API
1. Add support of Session Manager for Rest API;
2. Add Rest API for getting periodical report;
3. Add Rest API for getting console logs;
4. Fix KB-PROXY placement bug when topology file is provided;
5. Fix the logic for consolidating samples;
6. Refine Swagger definition files for Rest API;
7. Set the -c parameter in kb_gen_chart.py to required;

Known Issue / TODO:
1. The Pecan server logs will be dumped to clients;
2. The RestAPI "log" need remember the offset and only gets the
incremental

+++ kloudbuster/wrk_tool.py
@@ -129,10 +129,10 @@ class WrkTool(PerfTool):
     @staticmethod
     def consolidate_samples(results, vm_count):
         all_res = WrkTool.consolidate_results(results)
-        total_count = len(results) / vm_count
+        total_count = float(len(results)) / vm_count
         if not total_count:
             return all_res
 
-        all_res['http_rps'] = all_res['http_rps'] / total_count
-        all_res['http_throughput_kbytes'] = all_res['http_throughput_kbytes'] / total_count
+        all_res['http_rps'] = int(all_res['http_rps'] / total_count)
+        all_res['http_throughput_kbytes'] = int(all_res['http_throughput_kbytes'] / total_count)
         return all_res
